backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
openai==1.3.0
perplexity==0.0.1
supabase==2.0.0
chromadb==0.4.15
pydantic==2.5.0
python-dotenv==1.0.0
reportlab==4.0.7
apscheduler==3.10.4

